adding subject partially works right now
