+++ website/app.py
@@ -2,6 +2,7 @@ from flask import Flask, jsonify, render_template, request, send_file,redirect,u
 import json
 from .DBmanager import database_manager
 from flask_login import current_user, login_required
+import uuid
 
 from .auth import auth
 from . import create_app
@@ -26,8 +27,11 @@ def hello():
         return redirect("/login")
 
 @app.route('/profile_page')
+@login_required
 def profile_page(): 
-   return render_template("profile_page.html")
+   subjects = database_manager.getSubjects(current_user.user_id)
+   return render_template("profile_page.html", subjects=subjects)
+
 
 @app.route('/form', methods=['GET', 'POST'])
 def form():
@@ -49,6 +53,10 @@ def get_data():
 def get_js():
     return send_file('javascript/questionSubmitButtonPress.js', mimetype='application/javascript')
 
+# @app.route('/_subject')
+# def get_js():
+#     return send_file('javascript/subjectDisplayer.js', mimetype= 'text/javascript')
+
 @app.route('/new-post', methods=['POST'])
 def handle_new_post():
     data = request.form.to_dict()
@@ -63,3 +71,17 @@ def handle_new_post():
 def newPost(id,title, text, category):
     database_manager.addQuestion(id,title,text,category)
     pass 
+
+@app.route('/add_subject', methods=['GET', 'POST'])
+def add_subject():
+    # print(User.get_id(current_user))
+    user_id = current_user.user_id
+    print(database_manager.getTutorSubjects(user_id))
+    subjects = database_manager.getTutorSubjects(user_id)
+    if request.method == 'POST':
+        # Handle form submission for adding a subject
+        subject_name = request.form['subject']
+        database_manager.addSubject( user_id, subject_name, user_id, user_id)
+        subjects = database_manager.getTutorSubjects(user_id) # Update list of subjects
+    print(database_manager.getTutorSubjects(user_id))  # print the user's subjects
+    return render_template('profile_page.html', user=current_user, subjects=subjects)
